Reorganise concepts/explainability folder and rename to concepts/trustyai
