+++ frontend/src/api/trustyai/k8s.ts
@@ -5,7 +5,7 @@ import {
 } from '@openshift/dynamic-plugin-sdk-utils';
 import { K8sAPIOptions, K8sStatus, RouteKind, TrustyAIKind } from '~/k8sTypes';
 import { getRoute } from '~/api';
-import { TRUSTYAI_DEFINITION_NAME, TRUSTYAI_ROUTE_NAME } from '~/concepts/explainability/const';
+import { TRUSTYAI_DEFINITION_NAME, TRUSTYAI_ROUTE_NAME } from '~/concepts/trustyai/const';
 import { applyK8sAPIOptions } from '~/api/apiMergeUtils';
 import { TrustyAIApplicationsModel } from '~/api/models/trustyai';
