+++ frontend/src/concepts/trustyai/context/const.ts
@@ -1,0 +1,7 @@
+import { TrustyAIContextData } from '~/concepts/trustyai/context/types';
+
+export const DEFAULT_CONTEXT_DATA: TrustyAIContextData = {
+  refresh: () => Promise.resolve(),
+  biasMetricConfigs: [],
+  loaded: false,
+};
