+++ frontend/src/concepts/trustyai/context/types.ts
@@ -1,0 +1,8 @@
+import { BiasMetricConfig } from '~/concepts/trustyai/types';
+
+export type TrustyAIContextData = {
+  refresh: () => Promise<void>;
+  biasMetricConfigs: BiasMetricConfig[];
+  loaded: boolean;
+  error?: Error;
+};
