+++ frontend/src/concepts/trustyai/context/useDoesTrustyAICRExist.ts
@@ -1,10 +1,10 @@
 import React from 'react';
-import { ExplainabilityContext } from '~/concepts/explainability/ExplainabilityContext';
+import { TrustyAIContext } from '~/concepts/trustyai/context/TrustyAIContext';
 import { SupportedArea, useIsAreaAvailable } from '~/concepts/areas';
 
 const useDoesTrustyAICRExist = () => {
   const trustyAIAreaAvailable = useIsAreaAvailable(SupportedArea.TRUSTY_AI).status;
-  const { hasCR } = React.useContext(ExplainabilityContext);
+  const { hasCR } = React.useContext(TrustyAIContext);
 
   return [trustyAIAreaAvailable && hasCR];
 };
