+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationButton.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Button } from '@patternfly/react-core';
-import { useExplainabilityModelData } from '~/concepts/explainability/useExplainabilityModelData';
+import { useModelBiasData } from '~/concepts/trustyai/context/useModelBiasData';
 import { InferenceServiceKind } from '~/k8sTypes';
 import ManageBiasConfigurationModal from '~/pages/modelServing/screens/metrics/bias/biasConfigurationModal/ManageBiasConfigurationModal';
 
@@ -10,7 +10,7 @@ type BiasConfigurationButtonProps = {
 
 const BiasConfigurationButton: React.FC<BiasConfigurationButtonProps> = ({ inferenceService }) => {
   const [isOpen, setOpen] = React.useState(false);
-  const { biasMetricConfigs, loaded, refresh } = useExplainabilityModelData();
+  const { biasMetricConfigs, loaded, refresh } = useModelBiasData();
 
   React.useEffect(() => {
     if (loaded && biasMetricConfigs.length === 0) {
