+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationTable.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Button, ButtonVariant, ToolbarItem } from '@patternfly/react-core';
 import DashboardSearchField, { SearchType } from '~/concepts/dashboard/DashboardSearchField';
-import { BiasMetricConfig } from '~/concepts/explainability/types';
-import { useExplainabilityModelData } from '~/concepts/explainability/useExplainabilityModelData';
+import { BiasMetricConfig } from '~/concepts/trustyai/types';
+import { useModelBiasData } from '~/concepts/trustyai/context/useModelBiasData';
 import { InferenceServiceKind } from '~/k8sTypes';
 import DeleteBiasConfigurationModal from '~/pages/modelServing/screens/metrics/bias/biasConfigurationModal/DeleteBiasConfigurationModal';
 import DashboardEmptyTableView from '~/concepts/dashboard/DashboardEmptyTableView';
@@ -20,7 +20,7 @@ const BiasConfigurationTable: React.FC<BiasConfigurationTableProps> = ({
   inferenceService,
   onConfigure,
 }) => {
-  const { biasMetricConfigs, refresh } = useExplainabilityModelData();
+  const { biasMetricConfigs, refresh } = useModelBiasData();
   const [searchType, setSearchType] = React.useState<SearchType>(SearchType.NAME);
   const [search, setSearch] = React.useState('');
   const [cloneConfiguration, setCloneConfiguration] = React.useState<BiasMetricConfig>();
