+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationTableRow.tsx
@@ -6,7 +6,7 @@ import {
   DescriptionListGroup,
   DescriptionListTerm,
 } from '@patternfly/react-core';
-import { BiasMetricConfig } from '~/concepts/explainability/types';
+import { BiasMetricConfig } from '~/concepts/trustyai/types';
 
 type BiasConfigurationTableRowProps = {
   obj: BiasMetricConfig;
