+++ frontend/src/pages/modelServing/screens/metrics/bias/tableData.tsx
@@ -1,4 +1,4 @@
-import { BiasMetricConfig } from '~/concepts/explainability/types';
+import { BiasMetricConfig } from '~/concepts/trustyai/types';
 import { SortableData } from '~/components/table';
 
 export const columns: SortableData<BiasMetricConfig>[] = [
