+++ frontend/src/pages/modelServing/screens/metrics/bias/useBiasChartsBrowserStorage.ts
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { useBrowserStorage } from '~/components/browserStorage';
-import { BiasMetricConfig } from '~/concepts/explainability/types';
+import { BiasMetricConfig } from '~/concepts/trustyai/types';
 import { SetBrowserStorageHook } from '~/components/browserStorage/BrowserStorageContext';
 
 const SELECTED_CHARTS_STORAGE_KEY_PREFIX = 'odh.dashboard.xai.selected_bias_charts';
